% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_converter.R
\name{png_converter}
\alias{png_converter}
\title{Convert set of PNGs to PDF}
\usage{
png_converter(pngs, pdfname, extra.opts = "-density 300")
}
\arguments{
\item{pngs}{List of names (or with wildcards) of png files}

\item{pdfname}{Output PDF filename}

\item{extra.opts}{Options passed to \code{\link{im.convert}}}
}
\value{
Name of PDF output file
}
\description{
Converts PNGs to PDF
}

